+++ libsolidity/analysis/SyntaxChecker.cpp
@@ -237,8 +237,13 @@ bool SyntaxChecker::visit(FunctionDefinition const& _function)
 		if (v050)
 			m_errorReporter.syntaxError(_function.location(), "Functions without implementation cannot have modifiers.");
 		else
-			m_errorReporter.warning( _function.location(), "Modifiers of functions without implementation are ignored." );
+			m_errorReporter.warning(_function.location(), "Modifiers of functions without implementation are ignored." );
 	}
+	if (_function.name() == "constructor")
+		m_errorReporter.warning(_function.location(),
+			"This function is named \"constructor\" but is not the constructor of the contract. "
+			"If you intend this to be a constructor, use \"constructor(...) { ... }\" without the \"function\" keyword to define it."
+		);
 	return true;
 }